Full C# + Python AI Base Implementation, Revamp of Input and Logging in System, Base for Beginning of Gameplay loop

+++ Core/Engine/Behaviors/Coroutine.cs
@@ -9,7 +9,7 @@ namespace Automathon.Engine
     public class Coroutine : Behavior
     {
         public IEnumerator Enumerator;
-        private float waitTimer;
+        private int waitTimer;
         private Func<bool> pausedUntil;
 
         private Stack<IEnumerator> stack;
